requirements.txt
torch
torchvision
geopy
scikit-learn
numpy
matplotlib
fastapi
uvicorn[standard]
python-multipart
Pillow